MGMT-2466 Add migrations with gormigrate
The pattern I'm trying to establish here is a new file per migration.
This should allow us to organize well going forward. Each new migration
file should define a function matching the friendly name of the migration
with an id matching the timestamp of the migration.

To generate a timestamp I used `date -u +%Y%m%d%H%M%S`. It may be worth
creating a new make target to generate a new migration file with
the current timestamp so we don't have to validate correctness in
reviews.