AGENT-37 - Read ZTP CRs from filesystem to populate database
In the current POC of BILLI assisted-service is deployed
as a pod and kube-api is not available to it. The ZTP
CRDs is used as the API for users to define their cluster.

Since kube-api is not available here we've added a new
way to create clusters and infra-envs in the database by
having assisted-service read in ZTP CRs from the file-system.

Loading the CRs and calling existing bminventory functions
to create the cluster and infra env is initiated by an
init container. NOTE: It's temporarily embedded in assisted
service main.go.

Refactored params out of the create functions in clusterdeployment
and infraenv controller for code reuse between existing kube-api
and BILLI flows.